+++ lib/css_parser/parser.rb
@@ -164,6 +164,25 @@ module CssParser
       parse_block_into_rule_sets!(block, options)
     end
 
+    # Add a CSS rule by setting the +selectors+, +declarations+
+    # and +media_types+. Optional pass +filename+ , +offset+ for source
+    # reference too.
+    #
+    # +media_types+ can be a symbol or an array of symbols. default to :all
+    # optional fields for source location for source location
+    # +filename+ can be a string or uri pointing to the file or url location.
+    # +offset+ should be Range object representing the start and end byte locations where the rule was found in the file.
+    def append_rule!(selectors: nil, block: nil, filename: nil, offset: nil, media_types: :all)
+      rule_set = RuleSet.new(
+        selectors: selectors, block: block,
+        offset: offset, filename: filename
+      )
+
+      add_rule_set!(rule_set, media_types)
+    rescue ArgumentError => e
+      raise e if @options[:rule_set_exceptions]
+    end
+
     # Add a CSS rule by setting the +selectors+, +declarations+ and +media_types+.
     #
     # +media_types+ can be a symbol or an array of symbols. default to :all
@@ -172,10 +191,9 @@ module CssParser
     # +offset+ should be Range object representing the start and end byte locations where the rule was found in the file.
 
     def add_rule!(selectors, declarations, media_types = :all)
-      rule_set = RuleSet.new(selectors: selectors, block: declarations)
-      add_rule_set!(rule_set, media_types)
-    rescue ArgumentError => e
-      raise e if @options[:rule_set_exceptions]
+      warn '[DEPRECATION] `add_rule!` is deprecated. Please use `append_rule!` instead.', uplevel: 1
+
+      append_rule!(selectors: selectors, block: declarations, media_types: media_types)
     end
 
     # Add a CSS rule by setting the +selectors+, +declarations+, +filename+, +offset+ and +media_types+.
@@ -184,12 +202,11 @@ module CssParser
     # +offset+ should be Range object representing the start and end byte locations where the rule was found in the file.
     # +media_types+ can be a symbol or an array of symbols.
     def add_rule_with_offsets!(selectors, declarations, filename, offset, media_types = :all)
-      rule_set = RuleSet.new(
-        selectors: selectors, block: declarations,
-        offset: offset, filename: filename
+      warn '[DEPRECATION] `add_rule_with_offsets!` is deprecated. Please use `append_rule!` instead.', uplevel: 1
+      append_rule!(
+        selectors: selectors, block: declarations, media_types: media_types,
+        filename: filename, offset: offset
       )
-
-      add_rule_set!(rule_set, media_types)
     end
 
     # Add a CssParser RuleSet object.